SPLIT UP: Add GraqhQL API to create new food products & Android

+++ backend/apps/foods/models/food.py
@@ -40,7 +40,9 @@ class Food(Nutrients):
         default=UNIT_GRAM,
     )
 
-    weight = models.PositiveIntegerField(
+    weight = models.DecimalField(
+        max_digits=10,
+        decimal_places=2,
         default=100,
     )
